+++ maplibre_gl/test/helpers/fake_platform.dart
@@ -123,6 +123,16 @@ class FakeMapLibrePlatform extends MapLibrePlatform {
   Future<void> forceOnlineMode() async {}
 
   @override
+  Future<void> pauseMap() async {
+    calls.add(PlatformCall('pauseMap'));
+  }
+
+  @override
+  Future<void> resumeMap() async {
+    calls.add(PlatformCall('resumeMap'));
+  }
+
+  @override
   Future<bool> easeCamera(
     CameraUpdate cameraUpdate, {
     Duration? duration,